Add files via upload
Only 2 players (i don't know if i will add more in the future for now), can't change any settings (such as end of game) for now, but again i don't know if i will change that in the future, game ends when a player reaches 10 points, and is played in a terminal. As the repo description says, this is written in python.